requirements.txt
pydantic~=2.9.2
pydantic-settings~=2.6.1
numpy~=2.0.2
nltk~=3.9.1
pandas~=2.2.3
PyMuPDF 
Pillow~=11.0.0
spacy~=3.8.2
opencv-python~=4.10.0.84
langdetect~=1.0.9
scikit-learn~=1.5.2
transformers~=4.46.2
torch~=2.5.1
easyocr[ru]~=1.7.2
pymorphy2~=0.9.1
streamlit~=1.40.0
requests~=2.32.3
tqdm~=4.67.0